Dev pagina de productos (#8)
* Esto es una prueba

* prueba de seguridad 2

* prueba de seguridad 3

* Agregando clase productos en todas las capas

* Intento hacer funcionar el api de los productos

* Ya la pagina se crea apartir del sql

* Ya funciona el cargo de productos con todo y precios

* se arreglaron plugins del pom y se creo un api para la clase productos

* al darle click ya remite a una pagina en la que se muestra el producto y su valor

* añadiendo boton para carrito de compras

* Add files via upload

* Arreglo de clases

* Delete src/main/java/co/edu/eci/cvds/model/Usuario.java

* agregue bordes a los productos

* Add files via upload (#2)


* Se conectaron todas las pantallas de manera que el controlador del modulo privado funciona

* movi muchos archivos a las carpetas correctas e incorpore la pantalla solicitada por el cliente

* agregue mejores contenedores para los productos y un boton para añadirlos al carrito

* ya filtra por categorias y quite mucho js

* Filtra por categoria y hay botones pero no completamente terminados

* Dev front laura (#7)

* Dev pagina de productos (#6)

* Esto es una prueba

* prueba de seguridad 3

* Agregando clase productos en todas las capas

* Intento hacer funcionar el api de los productos

* Ya la pagina se crea apartir del sql

* Ya funciona el cargo de productos con todo y precios

* se arreglaron plugins del pom y se creo un api para la clase productos

* al darle click ya remite a una pagina en la que se muestra el producto y su valor

* añadiendo boton para carrito de compras

* Tomassuarez49 patch 1 (#1)

* Esto es una prueba

* prueba de seguridad 2

* Delete src/main/java/co/edu/eci/cvds/model/Usuario.java

---------


* Arreglo de clases

* Delete src/main/java/co/edu/eci/cvds/model/Usuario.java

* agregue bordes a los productos

* Add files via upload (#2)


* Dev pagina de productos (#3)

* Esto es una prueba

* prueba de seguridad 2

* prueba de seguridad 3

* Agregando clase productos en todas las capas

* Intento hacer funcionar el api de los productos

* Ya la pagina se crea apartir del sql

* Ya funciona el cargo de productos con todo y precios

* se arreglaron plugins del pom y se creo un api para la clase productos

* al darle click ya remite a una pagina en la que se muestra el producto y su valor

* añadiendo boton para carrito de compras

* Arreglo de clases

* Delete src/main/java/co/edu/eci/cvds/model/Usuario.java

* agregue bordes a los productos

* Add files via upload (#2)


---------


* Se conectaron todas las pantallas de manera que el controlador del modulo privado funciona

* Dev pagina de productos (#4)

* Esto es una prueba

* prueba de seguridad 2

* prueba de seguridad 3

* Agregando clase productos en todas las capas

* Intento hacer funcionar el api de los productos

* Ya la pagina se crea apartir del sql

* Ya funciona el cargo de productos con todo y precios

* se arreglaron plugins del pom y se creo un api para la clase productos

* al darle click ya remite a una pagina en la que se muestra el producto y su valor

* añadiendo boton para carrito de compras

* Arreglo de clases

* Delete src/main/java/co/edu/eci/cvds/model/Usuario.java

* agregue bordes a los productos

* Add files via upload (#2)


* Se conectaron todas las pantallas de manera que el controlador del modulo privado funciona

---------


* movi muchos archivos a las carpetas correctas e incorpore la pantalla solicitada por el cliente

* agregue mejores contenedores para los productos y un boton para añadirlos al carrito

* ya filtra por categorias y quite mucho js

* Filtra por categoria y hay botones pero no completamente terminados

---------


* Ya quedó el carrito conectado yeii

---------


* intentando resolver problemas

* aaa

* Ultimos cambios en controlador

---------

+++ src/main/java/co/edu/eci/cvds/model/Product.java
@@ -18,7 +18,7 @@ public class Product {
     @Column(name = "VALOR")
     private int valor;
     @Column(name = "DESCUENTO")
-    private int descuento;
+    private double descuento;
     @Column(name = "DESCRIPCIONBREVE")
     private String descripcionBreve;
     @Column(name = "DESCRIPCIONTECNICA")
@@ -31,4 +31,5 @@ public class Product {
     private String tipoUsuario;
     @Column(name = "ICONURL")
     private String iconUrl;
+
 }
